move the platformmethods thing to RemapInjectables, where its used
